Store DB in the app for shared reference
Just makes it easier to acces the database in one place. On app startup, the DB is loaded and then referenced throughout using Express's app.locals. This effectively changes nothng, since before this change, nodejs imports the db once, and then each additional import is ignored and the previous import is used from memory.

This diff does start to store active tokens as a map of token to user id. This allows a token to quickly checked against the object whether it exists or not and if it does, get the user id associated with it. From there, additional user details can be looked up on the user section of the DB as needed.