Setup Angular for further development
This commit initializes Angular in a `client` folder. Its dependencies are managed separately, but all within the same repository for an easy developer experience. From here, one can develop an Angular application that consumes the server's API.

`npm run dev` has been created to run both the server and the frontend content (webpack dev server) at once.

The frontend end code is configured to proxy /api to port 3000 /api, so that apis are accessible without CORS configuration.

No production serving of frontend content is configured. This is because there are multiple ways to serve the content, and one hasn't been chosen or even determined to be needed. This may never be hosted somewhere.

github actions have been configured to run backend tests on both the latest LTS and the current supported nodejs versions. Frontend tests are run using the latest LTS version only, since they are testing frontend content that runs in a browser, stable nodejs is needed, and runing on multiple versions would be pointless.